Fixes in fields in Contest. Also fixes in participate method in ContestParticipationServiceImpl and one new method in ContestParticipationServiceImpl, ContestParticipationRestController and ContestParticipationRepository. Two new classes: ContestScoreManager and RankManager. Some fixes in methods in JuryPhotoRatingServiceImpl and JuryPhotoRatingRepository and JuryPhotoRatingToJuryPhotoRatingOutputConverter.Also some changes in User and Photo classes.

+++ src/main/java/com/telerikacademy/web/photocontest/entities/Photo.java
@@ -40,7 +40,7 @@ public class Photo {
     private User user;
 
     @Column
-    private int total_score;
+    private Double total_score;
 
     @Column(name = "upload_date")
     private LocalDateTime createdAt;
@@ -55,6 +55,6 @@ public class Photo {
         this.contest = contest;
         this.createdAt = LocalDateTime.now();
         this.isActive = true;
-        this.total_score = 0;
+        this.total_score = 0.0;
     }
 }
